Change project structure
