--- SuccessDialog.axaml
@@ -1,32 +1,0 @@
-﻿<!--Copyright 2020-2021 dan0v
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.-->
-
-<Window xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d" d:DesignWidth="350" d:DesignHeight="150"
-        x:Class="xdelta3_cross_gui.SuccessDialog"
-        Title="Success"
-        DataContext="{Binding RelativeSource={RelativeSource Self}}"
-        Height="150" Width="350"
-        Icon="avares://xdelta3_cross_gui/Assets/img/Icon.png">
-  
-  <Grid RowDefinitions="*,Auto,Auto">
-    <TextBlock Grid.Row="0" HorizontalAlignment="Center" Text="Success" FontWeight="Bold" Margin="10"></TextBlock>
-    <TextBlock Grid.Row="1" TextWrapping="Wrap" Margin="10" TextAlignment="Center" Text="Your xDelta3 patches, or patch batches, have been successfuly created."></TextBlock>
-    <Button x:Name="btn_OpenDestination" Grid.Row="2" Content="Open output directory" HorizontalAlignment="Left" Margin="10" MinWidth="150"></Button>
-    <Button x:Name="btn_Dismiss" Grid.Row="2" Content="Dismiss" HorizontalAlignment="Right" Margin="10" Width="80"></Button>
-  </Grid>
-</Window>
