+++ Windows/Console.axaml
@@ -1,0 +1,33 @@
+﻿<!--Copyright 2020-2021 dan0v
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.-->
+
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        x:Class="xdelta3_cross_gui.Console"
+        Title="Console"
+        DataContext="{Binding RelativeSource={RelativeSource Self}}"
+        Height="500" Width="500"
+        Icon="avares://xdelta3_cross_gui/Assets/img/Icon.png">
+  
+  <Border BorderThickness="2,2,2,2" BorderBrush="Gray">
+    <Grid RowDefinitions="*">
+      <ScrollViewer Grid.Row="1" Background="Black" x:Name="sv_ScrollConsole" HorizontalScrollBarVisibility="Disabled" VerticalScrollBarVisibility="Auto">
+        <TextBlock x:Name="txt_blk_Output" Foreground="White" Margin="5" TextWrapping="Wrap"></TextBlock>
+      </ScrollViewer>
+    </Grid>
+  </Border>
+</Window>
