+++ Windows/Dialogs/ErrorDialog.axaml
@@ -1,0 +1,38 @@
+﻿<!--Copyright 2020-2021 dan0v
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.-->
+
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="500" d:DesignHeight="200"
+        x:Class="xdelta3_cross_gui.ErrorDialog"
+        Title="Error"
+        DataContext="{Binding RelativeSource={RelativeSource Self}}"
+        Height="200" Width="500"
+        Icon="avares://xdelta3_cross_gui/Assets/img/Icon.png">
+  
+  <Grid RowDefinitions="Auto,*,Auto,Auto">
+    <TextBlock Grid.Row="0" HorizontalAlignment="Center" Text="An error has occured" FontWeight="Bold" Margin="10"></TextBlock>
+    <Grid x:Name="grd_MissingFiles" Grid.Row="1" ColumnDefinitions="*,*" RowDefinitions="Auto,Auto,*" IsVisible="False">
+      <TextBlock HorizontalAlignment="Center" Grid.Row="0" Grid.ColumnSpan="2" Text="Missing Files"></TextBlock>
+      <TextBlock Grid.Column="0" Grid.Row="1" HorizontalAlignment="Center" Text="Old"></TextBlock>
+      <TextBlock Grid.Column="1" Grid.Row="1" HorizontalAlignment="Center" Text="New "></TextBlock>
+      <TextBlock TextWrapping="Wrap" x:Name="txt_blk_MissingOld" Grid.Column="0" Grid.Row="2" HorizontalAlignment="Center" Margin="5,0,2.5,0" ></TextBlock>
+      <TextBlock TextWrapping="Wrap" x:Name="txt_blk_MissingNew" Grid.Column="1" Grid.Row="2" HorizontalAlignment="Center" Margin="2.5,0,5,0" ></TextBlock>
+    </Grid>
+    <TextBlock Grid.Row="2" TextWrapping="Wrap" Margin="10" TextAlignment="Center" Text="Some options may have been reset. Please check before trying again."></TextBlock>
+    <Button x:Name="btn_Dismiss" Grid.Row="3" Content="Dismiss" HorizontalAlignment="Right" Margin="10" Width="80"></Button>
+  </Grid>
+</Window>
